reddit_scraper -> reddit
Moved all of reddit_scraper into reddit file inside of data-pipeline/scrapers. Additionally, added .env support in the file.
