+++ bin/wsdump
@@ -1,0 +1,8 @@
+#!/Users/dylanacosta/Desktop/Projects/market-sentiment/bin/python3
+# -*- coding: utf-8 -*-
+import re
+import sys
+from websocket._wsdump import main
+if __name__ == '__main__':
+    sys.argv[0] = re.sub(r'(-script\.pyw|\.exe)?$', '', sys.argv[0])
+    sys.exit(main())
