+++ requirements/requirements-data.txt
@@ -3,3 +3,4 @@ yfinance
 requests
 beautifulsoup4
 scipy
+praw
